+++ src/dom_tree.rs
@@ -1,5 +1,7 @@
 mod ops;
+mod traversal;
 mod tree;
 
 pub use ops::TreeNodeOps;
+pub use traversal::Traversal;
 pub use tree::Tree;